policy_pulse/pyproject.toml
[tool.poetry]
name = "policy_pulse"
version = "0.1.0"
description = "Simple AI policy/law summarizer"
authors = ["Ryan"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12.3"
crewai = "0.1.32"
unstructured = '==0.10.25'
pyowm = '3.3.0'
python-dotenv = "1.0.0"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignote = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
